refactor: modularize 3d dungeon view
