+++ src/games/dungeon-rpg-three/DungeonView.ts
@@ -1,11 +1,11 @@
 import * as THREE from 'three'
-import { ImprovedNoise } from 'three/examples/jsm/math/ImprovedNoise.js'
 import DungeonMap from '../dungeon-rpg/DungeonMap'
 import Player, { Direction } from '../dungeon-rpg/Player'
 import Hero from '../dungeon-rpg/Hero'
 import Enemy, { skeletonWarrior } from '../dungeon-rpg/Enemy'
-import armBox from '../../assets/arms/arm-box.json'
+import PlayerArms from './components/PlayerArms'
 import skeletonShape from '../../assets/enemies/json/skeleton-warrior.json'
+import { floorTexture, wallTexture, perlinTexture } from './utils/textures'
 
 export default class DungeonView3D {
   private scene: THREE.Scene
@@ -34,11 +34,7 @@ export default class DungeonView3D {
   private targetPos = new THREE.Vector3()
   private targetRot = 0
   private readonly animDuration = 200 // ms
-  private armsGroup?: THREE.Group
-  private leftUpper?: THREE.Mesh
-  private leftLower?: THREE.Mesh
-  private rightUpper?: THREE.Mesh
-  private rightLower?: THREE.Mesh
+  private arms: PlayerArms
   private readonly cellSize = 2
   private readonly wallNoiseScale = 25
 
@@ -112,7 +108,7 @@ export default class DungeonView3D {
     window.addEventListener('keydown', this.handleKeyDown)
 
     this.buildScene()
-    this.createArms()
+    this.arms = new PlayerArms(this.camera)
     // set initial camera state without animation
     this.camera.position.set(
       this.player.x * this.cellSize + this.cellSize / 2,
@@ -126,121 +122,9 @@ export default class DungeonView3D {
     this.targetRot = this.startRot
   }
 
-  private tileablePerlin(
-    noise: ImprovedNoise,
-    x: number,
-    y: number,
-    size: number,
-    scale: number,
-    z: number
-  ): number {
-    const u = x / size
-    const v = y / size
-    const w = size / scale
-    const nx = x / scale
-    const ny = y / scale
-    const n00 = noise.noise(nx, ny, z)
-    const n10 = noise.noise(nx - w, ny, z)
-    const n01 = noise.noise(nx, ny - w, z)
-    const n11 = noise.noise(nx - w, ny - w, z)
-    return n00 * (1 - u) * (1 - v) + n10 * u * (1 - v) + n01 * (1 - u) * v + n11 * u * v
-  }
-
-  private perlinTexture(
-    size = 256,
-    base = 30,
-    range = 50,
-    scale = 50
-  ) {
-    const canvas = document.createElement('canvas')
-    canvas.width = canvas.height = size
-    const ctx = canvas.getContext('2d') as CanvasRenderingContext2D
-    const imgData = ctx.createImageData(size, size)
-    const noise = new ImprovedNoise()
-    const z = Math.random() * 100
-    let i = 0
-    for (let y = 0; y < size; y++) {
-      for (let x = 0; x < size; x++) {
-        const v = this.tileablePerlin(noise, x, y, size, scale, z)
-        const c = Math.floor(base + (v + 1) * range)
-        imgData.data[i++] = c
-        imgData.data[i++] = c
-        imgData.data[i++] = c
-        imgData.data[i++] = 255
-      }
-    }
-    ctx.putImageData(imgData, 0, 0)
-    const tex = new THREE.CanvasTexture(canvas)
-    tex.wrapS = tex.wrapT = THREE.RepeatWrapping
-    return tex
-  }
-
-  private perlinTextureColor(
-    size = 256,
-    base: { r: number; g: number; b: number },
-    range: { r: number; g: number; b: number },
-    scale = 50
-  ) {
-    const canvas = document.createElement('canvas')
-    canvas.width = canvas.height = size
-    const ctx = canvas.getContext('2d') as CanvasRenderingContext2D
-    const imgData = ctx.createImageData(size, size)
-    const noise = new ImprovedNoise()
-    const z = Math.random() * 100
-    let i = 0
-    for (let y = 0; y < size; y++) {
-      for (let x = 0; x < size; x++) {
-        const v = this.tileablePerlin(noise, x, y, size, scale, z)
-        imgData.data[i++] = Math.floor(base.r + (v + 1) * range.r)
-        imgData.data[i++] = Math.floor(base.g + (v + 1) * range.g)
-        imgData.data[i++] = Math.floor(base.b + (v + 1) * range.b)
-        imgData.data[i++] = 255
-      }
-    }
-    ctx.putImageData(imgData, 0, 0)
-    const tex = new THREE.CanvasTexture(canvas)
-    tex.wrapS = tex.wrapT = THREE.RepeatWrapping
-    return tex
-  }
-
-  private floorTexture() {
-    return this.perlinTextureColor(
-      256,
-      { r: 80, g: 60, b: 40 },
-      { r: 30, g: 20, b: 15 },
-      40
-    )
-  }
-
-  private wallTexture() {
-    return this.perlinTextureColor(
-      256,
-      { r: 70, g: 70, b: 70 },
-      { r: 50, g: 50, b: 50 },
-      this.wallNoiseScale
-    )
-  }
-
-  private checkerTexture(color1: string, color2: string, squares = 8) {
-    const size = 64
-    const canvas = document.createElement('canvas')
-    canvas.width = size
-    canvas.height = size
-    const ctx = canvas.getContext('2d') as CanvasRenderingContext2D
-    const s = size / squares
-    for (let y = 0; y < squares; y++) {
-      for (let x = 0; x < squares; x++) {
-        ctx.fillStyle = (x + y) % 2 === 0 ? color1 : color2
-        ctx.fillRect(x * s, y * s, s, s)
-      }
-    }
-    const tex = new THREE.CanvasTexture(canvas)
-    tex.wrapS = tex.wrapT = THREE.RepeatWrapping
-    return tex
-  }
 
   private buildScene() {
-    const floorTex = this.floorTexture()
+    const floorTex = floorTexture()
     floorTex.repeat.set(
       this.map.width * this.cellSize,
       this.map.height * this.cellSize
@@ -262,7 +146,7 @@ export default class DungeonView3D {
     this.scene.add(floor)
 
 
-    const ceilTex = this.perlinTexture(256, 10, 20)
+    const ceilTex = perlinTexture(256, 10, 20)
 
     ceilTex.repeat.set(
       this.map.width * this.cellSize,
@@ -284,7 +168,7 @@ export default class DungeonView3D {
     )
     this.scene.add(ceiling)
 
-    const wallTex = this.wallTexture()
+    const wallTex = wallTexture(this.wallNoiseScale)
     wallTex.wrapS = wallTex.wrapT = THREE.RepeatWrapping
     const wallScale = this.wallNoiseScale
     for (let y = 0; y < this.map.height; y++) {
@@ -456,76 +340,8 @@ export default class DungeonView3D {
     })
   }
 
-  private createArms() {
-    this.armsGroup = new THREE.Group()
-
-    const vertsPer = 24
-    const idxPer = 36
-
-    const partGeo = (i: number) => {
-      const verts = armBox.vertices.slice(i * vertsPer * 3, (i + 1) * vertsPer * 3)
-      const idx = armBox.indices
-        .slice(i * idxPer, (i + 1) * idxPer)
-        .map((v) => v - i * vertsPer)
-      const g = new THREE.BufferGeometry()
-      g.setAttribute('position', new THREE.Float32BufferAttribute(verts, 3))
-      g.setIndex(idx)
-      g.computeVertexNormals()
-      return g
-    }
-
-    const upperGeo = partGeo(2)
-    const lowerGeo = partGeo(0)
-    const fistGeo = partGeo(1)
-
-    const armMat = new THREE.MeshBasicMaterial({ color: 0xdfcbbd })
-
-    const makeArm = () => {
-      const upper = new THREE.Mesh(upperGeo, armMat)
-      const lower = new THREE.Mesh(lowerGeo, armMat)
-      lower.position.y = -0.3
-      const fist = new THREE.Mesh(fistGeo, armMat)
-      fist.position.y = -0.14
-      lower.add(fist)
-      upper.add(lower)
-      return { upper, lower }
-    }
-
-    const left = makeArm()
-    const right = makeArm()
-
-    this.leftUpper = left.upper
-    this.leftLower = left.lower
-    this.rightUpper = right.upper
-    this.rightLower = right.lower
-
-    this.leftUpper.rotation.x = 2.08
-    this.leftUpper.rotation.z = 0.33
-    this.rightUpper.rotation.x = 2.08
-    this.rightUpper.rotation.z = -0.33
-    const scale = 1.2
-    this.leftUpper.scale.setScalar(scale)
-    this.leftLower.scale.setScalar(scale)
-    this.rightUpper.scale.setScalar(scale)
-    this.rightLower.scale.setScalar(scale)
-
-    this.leftUpper.position.set(-0.25, 0, -0.6)
-    this.rightUpper.position.set(0.25, 0, -0.6)
-
-    this.armsGroup.add(this.leftUpper)
-    this.armsGroup.add(this.rightUpper)
-    this.armsGroup.position.y = -0.45
-    this.camera.add(this.armsGroup)
-  }
-
   getArmSettings() {
-    return {
-      posY: this.armsGroup?.position.y ?? 0,
-      upperRotX: this.leftUpper?.rotation.x ?? 0,
-      lowerRotX: this.leftLower?.rotation.x ?? 0,
-      rotZ: this.leftUpper?.rotation.z ?? 0,
-      scale: this.leftUpper?.scale.x ?? 1,
-    }
+    return this.arms.getSettings()
   }
 
   updateArms(settings: {
@@ -535,28 +351,9 @@ export default class DungeonView3D {
     rotZ?: number
     scale?: number
   }) {
-    if (this.leftUpper && this.rightUpper && this.leftLower && this.rightLower && this.armsGroup) {
-      if (settings.posY !== undefined) this.armsGroup.position.y = settings.posY
-      if (settings.upperRotX !== undefined) {
-        this.leftUpper.rotation.x = settings.upperRotX
-        this.rightUpper.rotation.x = settings.upperRotX
-      }
-      if (settings.lowerRotX !== undefined) {
-        this.leftLower.rotation.x = settings.lowerRotX
-        this.rightLower.rotation.x = settings.lowerRotX
-      }
-      if (settings.rotZ !== undefined) {
-        this.leftUpper.rotation.z = settings.rotZ
-        this.rightUpper.rotation.z = -settings.rotZ
-      }
-      if (settings.scale !== undefined) {
-        this.leftUpper.scale.setScalar(settings.scale)
-        this.leftLower.scale.setScalar(settings.scale)
-        this.rightUpper.scale.setScalar(settings.scale)
-        this.rightLower.scale.setScalar(settings.scale)
-      }
-    }
+    this.arms.update(settings)
   }
+
 
   update() {
     if (this.torch) {
